Replace \{_} with _ during bib compilation
As a lot of tools, such as Mendeley, export the sources maintained by
them using \{_} to escape the underscore in URLs most of them were not
clickable in the source reference.

To fight this, the latex template now registers a source map which maps
the \{_} escape string back to the original _ during bibtex compilation.